Modify loading order of configuration
Configuration can throw an error when instantiated. This can't
be instantiated on module load or we need to catch and handle
an error on every module load, which defeats the purpose of
making it global.

This change loads the configuration and passes it to the trigger
class, allowing us to handle the error during the main function.

This change also includes a few minor incorrectly referenced
resource fixes.